Converted results to be floating point

+++ ex3.py
@@ -1,19 +1,19 @@
 print("I will now count my chickens:") #This line of code display the string 'I will now count my chickens:'
 
-print("Hens ", 25 + 30 / 6) #This line prints the string of Hens and then does the math. 1st it divides 30 by 6 and then adds it to 25 which displays 30
-print("Roosters ", 100 - 25 * 3 % 4) #This line displays the string Roosters and then does the math. It first multiplies 25 * 3 which gives 75 then does the Modulus of 75 by 4 whic
+print("Hens ", float(25 + 30 / 6)) #This line prints the string of Hens and then does the math. 1st it divides 30 by 6 and then adds it to 25 which displays 30
+print("Roosters ", float(100 - 25 * 3 % 4)) #This line displays the string Roosters and then does the math. It first multiplies 25 * 3 which gives 75 then does the Modulus of 75 by 4 whic
 # results in 3 and finally it subtracts 100 by 3 which displays 97
 
 print("Now I will count the eggs: ") #This line displays Now I will count the eggs: 
 
-print(3 + 2 + 1 - 5 + 4 % 2 - 1 / 4 + 6) #This line does the math. First it does the Mod which is 0 then it divides 1/4 which is .25 and finally it adds and subtracts which is 6.75
+print(float(3 + 2 + 1 - 5 + 4 % 2 - 1 / 4 + 6)) #This line does the math. First it does the Mod which is 0 then it divides 1/4 which is .25 and finally it adds and subtracts which is 6.75
 
 print("Is it true that 3 + 2 < 5 - 7? ") #This line prints the string Is it true that 3 + 2 < 5 - 7?
 
 print(3 + 2 < 5 - 7) #This line compares the result of 5 and -2 which displays False
 
-print("What is 3 + 2? ", 3 + 2) #This line displays the string and the result of 3 + 2
-print("What is 5 - 7? ", 5 - 7) #This line displays the string and the result of 5 - 7
+print("What is 3 + 2? ", float(3 + 2)) #This line displays the string and the result of 3 + 2
+print("What is 5 - 7? ", float(5 - 7)) #This line displays the string and the result of 5 - 7
 
 print("Oh, that's why it's False. ") #This line displays the string
